+++ src/Analysis/Engine/Impl/Infrastructure/Check.cs
