+++ src/Analysis/Engine/Impl/Infrastructure/Definitions/IDocumentReader.cs
