+++ src/Analysis/Engine/Impl/Infrastructure/DocumentReader.cs
