+++ src/Analysis/Engine/Impl/Infrastructure/Extensions/EnumerableExtensions.cs
