+++ src/Analysis/Engine/Impl/Infrastructure/Extensions/ExceptionExtensions.cs
