+++ src/Analysis/Engine/Impl/Infrastructure/Extensions/StringBuilderExtensions.cs
