+++ src/Analysis/Engine/Impl/Infrastructure/Extensions/StringExtensions.cs
