+++ src/Analysis/Engine/Impl/Infrastructure/LinearSpan.cs
