+++ src/Analysis/Engine/Impl/Infrastructure/SourceSpan.cs
