+++ src/Analysis/Engine/Impl/LanguageServer/Definitions/Structures.cs
